allow other weights downloads
